🎨 moved getDBUsers func to context
